chore: Move SelectNewMetricButton
